requirements/docs-requirements.txt
six
bottle
autobahn[twisted]
twisted
pyyaml
txaio
pydash
typing
requests
click
sphinx-click
